mltools: JSON: combine JSON.Int and JSON.Int64 into a single variant.
It was convenient to have these as separate variants when we were only
using this type to generate JSON.  However if we also use it to parse
JSON documents then integers in the document should only map to a
single variant.

+++ v2v/output_rhv_upload.ml
@@ -276,7 +276,7 @@ object
 
         let disk_size = ov.ov_virtual_size in
         let json_params =
-          ("disk_size", JSON.Int64 disk_size) :: json_params in
+          ("disk_size", JSON.Int disk_size) :: json_params in
 
         (* Ask the plugin to write the disk ID to a special file. *)
         let diskid_file = diskid_file_of_id id in
